+++ src/test/scala/sgit/localChangeRepo/AddCommandTest.scala
@@ -1,11 +1,11 @@
 package sgit.localChangeRepo
 
+import java.nio.file.{Files, Paths}
+
 import better.files._
 import org.scalatest._
 import sgit.createRepo.InitCommand
 import sgit.io._
-import sgit.localChangeRepo.AddCommand
-import sgit.objects.{Blob, StagedLine}
 
 class AddCommandTest extends FunSpec with BeforeAndAfter {
   before {
@@ -19,27 +19,28 @@ class AddCommandTest extends FunSpec with BeforeAndAfter {
       .createIfNotExists()
   }
   after {
-    ".sgit/".toFile.delete()
-    ".sgit/".toFile.parent + "/" + "READMES.md".toFile.delete()
-    ".sgit/".toFile.parent + "/" + "READMEBIS.md".toFile.delete()
+    if(".sgit/".toFile.exists) ".sgit".toFile.delete()
+    if("READMES.md".toFile.exists) ".sgit/".toFile.parent + "/" + "READMES.md".toFile.delete()
+    if("READMEBIS.md".toFile.exists) ".sgit/".toFile.parent + "/" + "READMEBIS.md".toFile.delete()
   }
 
-  describe("If the user write the command sgit add <args> in the sgir repository."){
-    it("The file(s) corresponding to the regex is/are added to the staged file.") {
-      val regexarg = List("[A-Za-z]*R[a-zA-Z]*.[a-z]*")
-
-      AddCommand.addAccordingTypeArg(regexarg)
-      //search the file corresponding to the regex in the user repo
-      val fileInUserRepo = RepoSearching.searchDirectoryFile("[A-Za-z]*R[a-zA-Z]*.[a-z]*".r)
-      val files = fileInUserRepo.filterNot(f => f.isDirectory)
-      //generate sha key to files
-      val shakeys = files.map(f => (f.sha1 +" "+ ".sgit/".toFile.parent.relativize(f).toString))
-      //retrieve the sha key in the file staged
-      val contentFile = ".sgit/staged".toFile.contentAsString.split("\r\n").toList
-      assert(contentFile == shakeys)
+  describe("If the user write the command sgit add <args> in the sgit repository."){
+    it("should correspond to the regex is/are added to the staged file.") {
+        val regexarg = List("[A-Za-z]*R[a-zA-Z]*.[a-z]*")
+        AddCommand.addAccordingTypeArg(regexarg)
+        //search the file corresponding to the regex in the user repo
+        val fileInUserRepo = RepoSearching.searchDirectoryFile("[A-Za-z]*R[a-zA-Z]*.[a-z]*".r)
+        val files = fileInUserRepo.filterNot(f => f.isDirectory)
+        //generate sha key to files
+        val shakeys = files.map(f => (f.sha1 + " " + ".sgit/".toFile.parent.relativize(f).toString))
+        //retrieve the sha key in the file staged
+        val contentFile = ".sgit/staged".toFile.contentAsString
+          .replace("\r", "")
+          .split("\n").toList
+        assert(contentFile == shakeys)
 
     }
-    it("The file(s) corresponding the the name(s) given in by the user is/are added into the staged file."){
+    it("should correspond to the name(s) given in by the user is/are added into the staged file."){
       val files :Seq[String] = Seq("READMES.md", "READMEBIS.md")
       AddCommand.addAccordingTypeArg(files)
 
@@ -49,10 +50,12 @@ class AddCommandTest extends FunSpec with BeforeAndAfter {
       val files2 = foundFile.filterNot(f => f.isDirectory)
       val shakeys = files2.map(f => (f.sha1 +" "+ ".sgit/".toFile.parent.relativize(f).toString))
       //retrieve the sha key in the file staged
-      val contentFile = ".sgit/staged".toFile.contentAsString.split("\r\n").toList
+      val contentFile = ".sgit/staged".toFile.contentAsString
+        .replace("\r", "")
+        .split("\n").toList
       assert(contentFile == shakeys)
     }
-    it("All the files in the repository are added the the staged file.") {
+    it(" should add all the files in the repository in the staged file.") {
       val args :Seq[String] = Seq(".")
       AddCommand.addAccordingTypeArg(args)
 
@@ -61,7 +64,9 @@ class AddCommandTest extends FunSpec with BeforeAndAfter {
       val files2 = files.filterNot(f => f.isDirectory)
       val shakeys = files2.map(f => (f.sha1 +" "+ ".sgit/".toFile.parent.relativize(f).toString))
       //retrieve the sha key in the file staged
-      val contentFile = ".sgit/staged".toFile.contentAsString.split("\r\n").toList
+      val contentFile = ".sgit/staged".toFile.contentAsString
+        .replace("\r", "")
+        .split("\n").toList
       assert(contentFile == shakeys)
     }
   }
